AS_develop: fixed few bugs. Add camera autosize.

+++ Assets/App/Scripts/Scenes/Level/Grid/BlockGridConfig.cs
@@ -7,7 +7,7 @@ namespace App.Scripts.Scenes.Level
     public class BlockGridConfig
     {
         public PoolData<BlockGridCell> BlockGridCellPoolData;
-        public int CellSize = 1;
+        public float CellSize = 1;
         public bool MirrorX = false;
         public bool MirrorZ = true;
     }
